Now-redundant code removed from memory.cpp.
Minor emendation of CellPool.cpp.

git-svn-id: http://projects.cs.kent.ac.uk/projects/cxxr/svn/trunk@84 549f2c98-2231-0410-b618-a09cd5e121c0

+++ src/main/CellPool.cpp
@@ -71,7 +71,7 @@ void CellPool::checkCell(const void* p) const
     bool found = false;
     for (vector<void*>::const_iterator it = m_superblocks.begin();
 	 !found && it != m_superblocks.end(); ++it) {
-	long offset = pc - reinterpret_cast<const char*>(*it);
+	ptrdiff_t offset = pc - reinterpret_cast<const char*>(*it);
 	if (offset >= 0 && offset < static_cast<long>(m_superblocksize)) {
 	    found = true;
 	    if (offset%m_cellsize != 0) {
